+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/ESTLocationBeaconBulkUpdater.h
@@ -15,28 +15,28 @@
 
 #define ESTLocationBeaconBulkUpdaterErrorDomain @"ESTLocationBeaconBulkUpdaterErrorDomain"
 
-/*
+/**
  *  Defines types of errors that can occur inside the bulk updater mechanis.
  */
 typedef NS_ENUM(NSInteger, ESTLocationBeaconBulkUpdaterError)
 {
-    /*
+    /**
      *  Device passed as a part of bulk update was not in range.
      */
     ESTLocationBeaconBulkUpdaterErrorDeviceDiscoveryFailed
 };
 
-/*
+/**
  *  Defines the possible statuses of the whole bulk updater mechanism.
  */
 typedef NS_ENUM(NSInteger, ESTBulkUpdaterStatus)
 {
-    /*
+    /**
      * Bulk updater is ready to `startWithUpdateConfigurations:`.
      */
     ESTBulkUpdaterStatusIdle = 0,
     
-    /*
+    /**
      *  Bulk updater is performing bulk update actions. Cancel bulk before using `startWithUpdateConfigurations:` again.
      */
     ESTBulkUpdaterStatusRunning
@@ -47,27 +47,27 @@ typedef NS_ENUM(NSInteger, ESTBulkUpdaterStatus)
  */
 typedef NS_ENUM(NSInteger, ESTBulkUpdaterDeviceUpdateStatus)
 {
-    /*
+    /**
      * Not able to determine current update status.
      */
     ESTBulkUpdaterDeviceUpdateStatusUnknown,
-    /*
+    /**
      * Device not detected yet. Scanning in progress.
      */
     ESTBulkUpdaterDeviceUpdateStatusScanning,
-    /*
+    /**
      * Device detected. Waiting to connect and update settings.
      */
     ESTBulkUpdaterDeviceUpdateStatusPendingUpdate,
-    /*
+    /**
      * Device is beeing connected to and updated.
      */
     ESTBulkUpdaterDeviceUpdateStatusUpdating,
-    /*
+    /**
      * Device update succeeded.
      */
     ESTBulkUpdaterDeviceUpdateStatusSucceeded,
-    /*
+    /**
      * Device update failed.
      */
     ESTBulkUpdaterDeviceUpdateStatusFailed